refactor : 학교 검색 결과 페이지에서 훅 분리
